Embellish ill-fated test suite
