+++ api/ombpdf/js/prosemirror-fun/test.js
@@ -2,102 +2,41 @@
 // don't have a spare day to figure out how to set up Jest with
 // Babel.
 
-import assert from 'assert';
-import {Node} from "prosemirror-model";
+import * as convertDocTests from './test-convert-doc';
 
-import convertDoc from './convert-doc';
-import schema from './policy-schema';
+const TESTS = [
+  [convertDocTests, './test-convert-doc'],
+];
 
-function makeDbDoc(children) {
-  return {node_type: 'policy', children: children};
+function collectTests() {
+  const tests = {};
+
+  TESTS.forEach(args => {
+    const [moduleTests, prefix] = args;
+    Object.keys(moduleTests).forEach(name => {
+      tests[`${prefix}/${name}`] = moduleTests[name];
+    });
+  });
+
+  return tests;
 }
-
-function captureWarnings(fn) {
-  const oldWarn = console.warn;
-  const warnings = [];
-
-  console.warn = msg => warnings.push(msg);
-
-  try {
-    fn();
-  } finally {
-    console.warn = oldWarn;
-  }
-
-  return warnings;
-}
-
-const TESTS = {
-  testConvertDocChecksDoc() {
-    assert.throws(() => {
-      convertDoc({children: []});
-    }, /Invalid content for node doc/);
-  },
-  testEmptySectionFails() {
-    assert.throws(() => {
-      Node.fromJSON(schema, {
-        type: 'doc',
-        content: [{type: 'section', content: []}],
-      }).check();
-    }, /Invalid content for node section/);
-  },
-  testUnimplementedChildWorks() {
-    let doc;
-    const warnings = captureWarnings(() => {
-      doc = convertDoc(makeDbDoc([{
-        node_type: 'sec',
-        children: [{node_type: 'boof', foo: 'barf'}],
-      }]));
-    });
-
-    assert.deepEqual(warnings, ["Unknown node_type: boof"]);
-
-    const child = doc.firstChild.firstChild;
-    assert.deepEqual(JSON.parse(child.attrs.data), {
-      node_type: 'boof',
-      foo: 'barf'
-    });
-    assert.equal(child.type, schema.nodes.unimplemented_child);
-  },
-  testUnimplementedContentWorks() {
-    let doc;
-    const warnings = captureWarnings(() => {
-      doc = convertDoc(makeDbDoc([{
-        node_type: 'sec',
-        children: [
-          {
-            node_type: 'para',
-            content: [{content_type: 'blarg', foo: 'bar'}],
-            children: [],
-          }
-        ],
-      }]));
-    });
-
-    assert.deepEqual(warnings, ["Unknown content_type: blarg"]);
-
-    const child = doc.firstChild.firstChild.firstChild;
-    assert.deepEqual(JSON.parse(child.attrs.data), {
-      content_type: 'blarg',
-      foo: 'bar'
-    });
-    assert.equal(child.type, schema.nodes.unimplemented_content);
-  },
-};
 
 export default function runTests() {
   let currTest;
+  const tests = collectTests();
+
+  let testCount = Object.keys(tests).length;
 
   console.groupCollapsed('Running tests.');
 
   try {
-    Object.keys(TESTS).forEach(name => {
+    Object.keys(tests).forEach(name => {
       currTest = name;
       console.log(`Running test ${name}...`);
-      TESTS[name]();
+      tests[name]();
     });
     console.groupEnd();
-    console.log(`All ${Object.keys(TESTS).length} tests passed.`);
+    console.log(`All ${testCount} tests passed.`);
   } catch (e) {
     console.groupEnd();
     console.error(currTest, 'failed with exception', e);
